Module file
Module for overriding the form - there are different categories on site. One category is "linia". There are different images associated with different "linia". There are 3 lines of product. With every line there is appropriate picture associated with it, which is shown on node page. User wanted that, when he/she creates the node, and declare the linia, appropriate picture to be automatically associated with it. That is achieved with this module.